--- Commands/DataBase/CommandInterface.php
@@ -1,20 +1,0 @@
-<?php
-
-
-namespace GrapheneNodeClient\Commands\DataBase;
-
-use GrapheneNodeClient\Commands\CommandQueryDataInterface;
-
-interface CommandInterface
-{
-    /**
-     * @param CommandQueryDataInterface $commandQueryData
-     * @return mixed
-     */
-    public function execute(CommandQueryDataInterface $commandQueryData);
-
-    /**
-     * @return array
-     */
-    public function getQueryDataMap();
-}
