--- Commands/DataBase/GetActiveVotesCommand.php
@@ -1,17 +1,0 @@
-<?php
-
-
-namespace GrapheneNodeClient\Commands\DataBase;
-
-
-class GetActiveVotesCommand extends CommandAbstract
-{
-    /** @var string */
-    protected $method = 'get_active_votes';
-
-    /** @var array */
-    protected $queryDataMap = [
-        '0' => ['string'], //author
-        '1' => ['string'], //permlink
-    ];
-}
