+++ Commands/Single/GetAccountCountCommand.php
@@ -1,14 +1,11 @@
 <?php
 
 
-namespace GrapheneNodeClient\Commands\DataBase;
+namespace GrapheneNodeClient\Commands\Single;
 
 
 class GetAccountCountCommand extends CommandAbstract
 {
     /** @var string */
     protected $method = 'get_account_count';
-
-    /** @var array */
-    protected $queryDataMap = [];
 }
