+++ Commands/Single/GetAccountHistoryCommand.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetAccountHistoryCommand extends CommandAbstract
+{
+    /** @var string */
+    protected $method = 'get_account_history';
+}
