+++ Commands/Single/GetAccountVotesCommand.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetAccountVotesCommand extends CommandAbstract
+{
+    /** @var string */
+    protected $method = 'get_account_votes';
+}
