+++ Commands/Single/GetActiveVotesCommand.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetActiveVotesCommand extends CommandAbstract
+{
+    /** @var string */
+    protected $method = 'get_active_votes';
+}
