+++ Commands/Single/GetApiByNameCommand.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+/**
+ * ONLY for STEEM
+ */
+class GetApiByNameCommand extends CommandAbstract
+{
+    protected $method       = 'get_api_by_name';
+}
