+++ Commands/Single/GetBlockCommand.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetBlockCommand extends CommandAbstract
+{
+    /** @var string */
+    protected $method = 'get_block';
+}
