+++ Commands/Single/GetContentCommand.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetContentCommand extends CommandAbstract
+{
+    /** @var string */
+    protected $method = 'get_content';
+}
