+++ Commands/Single/GetContentRepliesCommand.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetContentRepliesCommand extends CommandAbstract
+{
+    /** @var string */
+    protected $method = 'get_content_replies';
+}
