+++ Commands/Single/GetCurrentMedianHistoryPriceCommand.php
@@ -1,14 +1,11 @@
 <?php
 
 
-namespace GrapheneNodeClient\Commands\DataBase;
+namespace GrapheneNodeClient\Commands\Single;
 
 
 class GetCurrentMedianHistoryPriceCommand extends CommandAbstract
 {
     /** @var string */
     protected $method = 'get_current_median_history_price';
-
-    /** @var array */
-    protected $queryDataMap = [];
 }
