+++ Commands/Single/GetDiscussionsByAuthorBeforeDateCommand.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetDiscussionsByAuthorBeforeDateCommand extends CommandAbstract
+{
+    protected $method         = 'get_discussions_by_author_before_date';
+}
