+++ Commands/Single/GetDiscussionsByBlogCommand.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetDiscussionsByBlogCommand extends CommandAbstract
+{
+    protected $method            = 'get_discussions_by_blog';
+}
