+++ Commands/Single/GetDiscussionsByFeedCommand.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetDiscussionsByFeedCommand extends CommandAbstract
+{
+    protected $method            = 'get_discussions_by_feed';
+}
