+++ Commands/Single/GetDiscussionsByTrendingCommand.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetDiscussionsByTrendingCommand extends CommandAbstract
+{
+    protected $method       = 'get_discussions_by_trending';
+}
