+++ Commands/Single/GetDynamicGlobalPropertiesCommand.php
@@ -1,14 +1,11 @@
 <?php
 
 
-namespace GrapheneNodeClient\Commands\DataBase;
+namespace GrapheneNodeClient\Commands\Single;
 
 
 class GetDynamicGlobalPropertiesCommand extends CommandAbstract
 {
     /** @var string */
     protected $method = 'get_dynamic_global_properties';
-
-    /** @var array */
-    protected $queryDataMap = [];
 }
