+++ Commands/Single/GetOpsInBlock.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+class GetOpsInBlock extends CommandAbstract
+{
+    /** @var string */
+    protected $method = 'get_ops_in_block';
+}
