+++ Commands/Single/GetTrendingCategoriesCommand.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetTrendingCategoriesCommand extends CommandAbstract
+{
+    protected $method = 'get_trending_categories';
+}
