+++ Commands/Single/GetWitnessesByVoteCommand.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetWitnessesByVoteCommand extends CommandAbstract
+{
+    protected $method       = 'get_witnesses_by_vote';
+}
