+++ Commands/Single/LoginCommand.php
@@ -1,21 +1,14 @@
 <?php
 
 
-namespace GrapheneNodeClient\Commands\Login;
+namespace GrapheneNodeClient\Commands\Single;
 
 /**
- * Class LoginCommand
+ * ONLY for STEEM
  * This must be called prior to requesting other APIs. Other APIs may not be accessible until the client has
  * sucessfully authenticated.
- *
- *
- * @package GrapheneNodeClient\Commands\Login
  */
 class LoginCommand extends CommandAbstract
 {
     protected $method       = 'login';
-    protected $queryDataMap = [
-        0 => ['string'],
-        1 => ['string']
-    ];
 }
